Week 13

+++ app/Http/Controllers/SupplierController.php
@@ -91,6 +91,7 @@ class SupplierController extends Controller
      */
     public function destroy(Supplier $supplier)
     {
+        $this->authorize('delete-permission',$supplier);
         try{
             $supplier->delete();
             return redirect()->route('suppliers.index')->with('status','Data Supplier berhasil dhapus');
